+++ Dashboard/SimuMan/serverdistance.h
@@ -19,7 +19,7 @@ public slots:
     void bytesWritten(qint64);
 
 
-    void update(int);
+    void update(float);
