+++ Dashboard/SimuMan/servertangage.cpp
@@ -37,11 +37,11 @@ void servertangage::bytesWritten(qint64 nb){
      qDebug() << "" << nb << " Bytes sent ..." ;
 }
 
-void servertangage::update(int newval)
+void servertangage::update(float newval)
 {
     if (isconnected==true){
 
-        float f = (float)newval;
+        float f = newval;
          //qDebug() <<f;
         QByteArray x(reinterpret_cast<const char *>(&f), sizeof(f)) ;
         a=x;
