+++ Dashboard/SimuMan/servervitesse.cpp
@@ -38,13 +38,13 @@ void servervitesse::bytesWritten(qint64 nb){
      qDebug() << "" << nb << " Bytes sent ..." ;
 }
 
-void servervitesse::update(int newval)
+void servervitesse::update(float newval)
 {
 
 
     if (isconnected==true){
 
-        float f = (float)newval;
+        float f = newval;
          //qDebug() <<f;
         QByteArray x(reinterpret_cast<const char *>(&f), sizeof(f)) ;
         a=x;
